+++ Nur.Domain/Entities/Products/Product.cs
@@ -1,6 +1,6 @@
 ﻿using Nur.Domain.Commons;
-using Nur.Domain.Entities.Attachments;
 using Nur.Domain.Enums;
+using Nur.Domain.Entities.Attachments;
 
 namespace Nur.Domain.Entities.Products;
 
@@ -8,7 +8,7 @@ public class Product : Auditable
 {
     public string Name { get; set; }
     public decimal Price { get; set; }
-    public long Quantity { get; set; }
+    public double Quantity { get; set; }
     public string Description { get; set; }
     public Unit Unit { get; set; }
